Add cybozu.validate package
cybozu.validate is a normalization and validation system for
Protocol Buffers.

The repository is set up for a Buf module in order to prepare
publishing protobuf files on BSR.

This commit does not include any code generator but includes
a hand-written Go code `examples/validation_cybozu_validate2.pb.go`
to demonstrate what to be generated.