Render Distinct Errors for Validation Failures
As a caller of `cybozu.validate` I want to take specific actions when
certain validations errors occur.

This changeset will render a distinct for every different type of
validation that is configured for a given message. As a caller you will
be able to use `errors.Is` to check for a specific validation error in
the returned error chain.